+++ src/main/java/com/citius/patientms/service/DemographicService.java
@@ -9,7 +9,7 @@ public interface DemographicService {
 
 	Demographic saveDemographic(Demographic demo);
 
-	Demographic updateDemographic(long id, Demographic demographic);
+	Demographic updateDemographic(int id, Demographic demographic);
 
 	List<Demographic> getAll();
 	public Demographic dtoTOEntity(DemographicDto dto);
